Add terms and conditions checkbox to contat form
