+++ enrolment/static/sass/company-profile-details.scss
@@ -381,15 +381,15 @@
   .btn-container {
     margin-top: 8px;
     margin-bottom: 0;
-    .button {
-      border: 2px solid #004f95;
-      color: #004f95;
-      padding: 15px;
-      text-align: center;
-      font-weight: bold;
-      display: block;
-      background: white;
-    }
+  }
+  .button {
+    border: 2px solid #004f95;
+    color: #004f95;
+    padding: 15px;
+    text-align: center;
+    font-weight: bold;
+    display: block;
+    background: white;
   }
 }
