+++ enrolment/static/sass/main.scss
@@ -330,28 +330,29 @@ label, input, button, select, textarea {
 .btn-container {
   margin-bottom: 30px;
   margin-top: 30px;
-  .button {
-    background-color: #00549F;
-    position: relative;
-    display: -moz-inline-stack;
-    display: inline-block;
-    padding: 21px 24px;
-    border: none;
-    border-radius: 0;
-    -webkit-appearance: none;
-    font-weight: 500;
-    text-decoration: none;
-    -webkit-font-smoothing: antialiased;
-    cursor: pointer;
-    color: #fff;
-    box-sizing: border-box;
-    vertical-align: top;
-    line-height: 1;
-    font-family: 'Roboto', sans-serif;
-    &:focus {
-      background: #007EA3;
-      outline: 0;
-    }
+}
+
+.button {
+  background-color: #00549F;
+  position: relative;
+  display: -moz-inline-stack;
+  display: inline-block;
+  padding: 21px 24px;
+  border: none;
+  border-radius: 0;
+  -webkit-appearance: none;
+  font-weight: 500;
+  text-decoration: none;
+  -webkit-font-smoothing: antialiased;
+  cursor: pointer;
+  color: #fff;
+  box-sizing: border-box;
+  vertical-align: top;
+  line-height: 1;
+  font-family: 'Roboto', sans-serif;
+  &:focus {
+    background: #007EA3;
+    outline: 0;
   }
 }
 
@@ -923,7 +924,9 @@ a.ed-filled-button-primary {
 }
 
 .ed-contact-form-container {
-  margin-bottom: 40px;
+  .errorlist.nonfield {
+    display: none;
+  }
   .ed-contact-form-inner {
     border: solid 1px #d8d8d8;
     border-bottom: 0;
@@ -931,8 +934,7 @@ a.ed-filled-button-primary {
   li {
     margin-bottom: 20px;
   }
-  .btn-container .button {
-    padding: inherit;
+  .button {
     width: 100px;
     font-size: 18px;
     border-radius: 3px;
